Add state string creation and testing useSearchParams
The state string creation is so we can move the state as high as needed from the Login component.
Later on, the useEffect may be changed to handle user information rather than just a state string.
In addition, the useSearchParams hook will be needed when redirected from the Discord OAuth authorization,
however it may be moved to a separate component later on