Change localStorage token key value pair
The random string generated is stored as the value for the key of 'token'.
Remove the need to pass the value of the string as a prop to Login component since it can just access
localStorage directly.
In addition, resolves the issue of resetting the csrfToken value when remounting after Discord redirect.